LayoutTests/fast/scrolling/mac/keyboard-scrolling-with-mouse-scroll-expected.txt

PASS endY >= startY is true
PASS successfullyParsed is true

TEST COMPLETE
fixed
